Added skeleton webgl support, switched to synchronous load of plugins specified in a json file.
Thus the set of plugins is now completely data-driven and new plugins can
be added to the total set without any modification of the core code.

The webgl support has only been tested on FF7/8 on Linux, but should work
the same on any browser / OS, with the possible exception of Opera.