Added two new plugins: Accumulate which accumulates an incomming float value. Makes the 'delta time' plugin useful, since the two can now be used in conjection to implement smooth time scaling. The other new plugin is a color blend modulator taking two colors and a linear mix coefficient between 0 and 1.
Fixed the camera viewport issues.

Rewrote the pipeline update logic and removed a lot of cruft. The core
design is now much nicer, and the code a little faster. Quite a few plugins
have been updated to fit with this new regime. The payoff is that a few
persistent corner case flow quicks have finally been fixed.

The texture implementation now check for and warns about non 2^nx2^m
textures, which are not usable from WebGL. These currently no automatic
rescaling on load.