Added new experimental plugins: Audio player and Audio from URL. Tested on FF 11, doesn't work in Chome v14 on Linux (no Audio API). I'll test with v18 or on Windows later. Let me know if it works for you and on what browser / OS combination.
The audio player receives two sources: An audio source from an Audio from
URL instance and a bool instructing the player to start or stop playback
(hook up a toggle button to test). If audio is playing when the graph
playback is stopped, it is stopped also. There's currently no provision
for pausing when graph playback is paused.

The browser wars have necessitated a quirk: When specifying an audio source
URL, there's no need to specify a file extension - if one is entered, it is
stripped. An appropriate extension is auto-selected based on browser
capabilities. MP4 AAC, OGG Vorbis, MP3 and PCM Wave are tried in that
order, and so samples must be uploaded as a .mp4, .ogg, .mp3, .wav set.

I've provided a test sample set in ".../data/audio/track" you can use for
testing.