Updated the mesh implementation to calculate per face normals if the model doesn't have any on load. Updated apply_uniforms to get the current mesh as a parameter and updated all the shader implementations to conform.
Furthermore, the implementation of the from mesh custom, from mesh and
texture diffuse shaders has been updated to use the diffuse texture
specified my the mesh in none is attached as an overload. If the current
mesh does not specific a diffuse texture the old behavior of disabling
texturing is used as a fallback.