Fixed the graph update to allow for selective dataflow though nested graphs and other plugins that might want to take advantage of this new addition to the plugin API.
As new method has been added: query_output(slot). If left out of a given
plugin implementation it will function as normal, but if implemented,
returning true will signal the the specified slot has changed this update
and false that it has not (which will inhibit dataflow to any connected
clients).