Core update: Stopped using setInterval in favor of requestAnimationFrame.
Besides making sure that rendering stops when the app is not in a visible
browser tab, this switch to a newer API also eliminated timer / frame
compositing aliasing. On my client it only improves rendering performance
a little, but it does seem to have a positive effect on UI responsiveness
during playback.