Renderer: Complete revamp of ComposeShader, scene rendering, mesh rendering, shader caching and material capability hash generation.
With this change we can now support dynamic, cached, lazily evaluated
shader recompilation in response to changes of all relevant material
properties, including all texture map types. In order to support this
change the scene renderer has been changed to no longer accept an
externally supplied shader. Instead, an overload material can now be
supplied directly. If an external material is specified, it will take
preference over the individual mesh materials, but if no texture map of
a given type has been specified in the external material, it will be
taken from the mesh material if such a map is specified there.

This gets rid of the contrived need for doing a quad -> from mesh shader
to make material changes to a whole scene. To support applying custom
shaders to whole scenes, a custom scene renderer plugin will be added
later.