Refactored the remaning plugins to use prototype: url audio, url scene, url texture, xyz -> vector, vector ... add, ... display, ... magnitude, ... multiply, ... normalize, ... scale, ... transform and xor plugins.
This pass also uncovered the following two bugs which were fixed:

* Vector magnitude's internal value for the default magnitude was
initialized to vector, not a float. This would have caused a crash if the
plugin had been attached without any inputs.

* Vector normalize: Wrong logic led to a division by zero with mag == 0.0,
with resulted in the emission of a vector comprised of NaNs. This case
now emits the vector [0, 0, 0] as one would expect.

The HTML update of vector display was optimized by using .text() instead
of .html().