From mesh custom shader: Bugfix: When a new dynamic input slot was created (or loaded) the internal state of the data for that slot was set to null.
This was a problem when shader inputs were created, but left unconnected.

Instead, the default shader input value is now the default value for its
datatype, just like with proxies and registers. One can now assume that
the value will always be something reasonable, even when the slot isn't
actully connected to anything.