As a first demonstration of our ability to add external JS projects, I've redesigned the UI of the from mesh custom shader plugin to feature a much nicer shader editor (ACE, the required files are included in this commit).
The rest of the UI has been touched up to give a larger code editor,
clearer dialog design, better button placement and labels, explanatory
region labels and fixes to the style of the popup dialog for
slot name input.

The main script is added to the document once if the current graph contains
one or more custom shader plugins. Language mode and theme are loaded once
on first use.