Various core improvements:
* The logic for load plugin state objects has been improved so as to make it
more robust against updates in the plugin state template. Now, a value
will only be loaded from the sequence if the members already exists. Thus,
obsolete values are skipped and will not propagate into new sequences or
into new sequences from old snippets. It also makes it possible to
introduce new state values with default values defined by the plugin
implementations.

* Add logic to ensure dialogs close when OK is clicked.

* Add core check for flag 'always_update' stored in the state of any
plugin that declares itself 'e2_is_graph'. If this is set to false,
the plugin will only be updated if the input context changes.

Both the graph and loop plugin has been updated with a options dialog. Both
plugins can be set to update always or conditionally based on whether
inputs change, and the graph  additionally allows configuration of the
render target width and height in the range 2 - 4096 pixels.