Updated the logic of array generators (array from size / array from object member) to decorate their emitted ArrayBuffers with the original datatype and stride, to allow dependent plugins to operate coherently on supplied arrays. Other changes relating to this update included here are:
* Since the datatype can now be inferred from the supplied array, new get
/ set plugins have been made to replace the existing ones. They are simpler
and do not require the user to specify the data type manually. The old
plugins are still available but are now called "Get as" and "Set as".
These are still useful to perform binary data conversion.

* The array length plugin now emits the length in instances, not the length
in bytes as before. Much more intuitive and convenient to work with, but
does require any existing snippets to be updated.

* Updated plugin icons.