Changed the logic of the tooltip implementation to build the text dynamically on mouse over rather than statically when the DOM for the plugin is created. This has the following adantages:
* Only a compact identifier is now stored in DOM alt="" tags rather than the whole text. The identifier has the form <node uid as int> for nodes and <node uid as int>_[s|o][i|o]_<slot index|slot uid> for slots. This is used in conjunction with the newly added graph capability for fast lookup by uid to dynamically resolve the required data on the target object to create the info text dynamically in onShowTooltip().

That saves a lot of redundant strings lying around in the DOM and fixes the issue where dynamic or static slots declared as ANY was still shown as 'Arbitrary' in info, even though its type had changed dynamically. The info view now correctly reflects the actual data type of a given slot.